lint with Solc parser

+++ contracts/ConvertLib.sol
@@ -1,12 +1,12 @@
 // SPDX-License-Identifier: MIT
 
-pragma solidity = 0.7.3;
+pragma solidity =0.7.3;
 
 library ConvertLib {
-    function convert(uint amount, uint conversionRate)
+    function convert(uint256 amount, uint256 conversionRate)
         public
         pure
-        returns (uint convertedAmount)
+        returns (uint256 convertedAmount)
     {
         return amount * conversionRate;
     }
