+++ libraries/ArrayUtils.sol
@@ -1,23 +1,24 @@
 // SPDX-License-Identifier: MIT
 
-pragma solidity = 0.7.3;
+pragma solidity =0.7.3;
 
 /** Array wraper
-* min() - returns minimal array element
-* max() - returns maximal array element
-* sum() - returns sum of all array elements
-* set(uint []) - set array
-* get() - returns stored array
-* sort() - sorts all array elements
-*/
+ * min() - returns minimal array element
+ * max() - returns maximal array element
+ * sum() - returns sum of all array elements
+ * set(uint []) - set array
+ * get() - returns stored array
+ * sort() - sorts all array elements
+ */
 
 //Uint Array
 
 contract UintUtils {
-    uint[] private data;
-    function UintArray(uint[] memory _data) public {
-        data = new uint[](_data.length);
-        for (uint i = 0; i < _data.length; i++) {
+    uint256[] private data;
+
+    function UintArray(uint256[] memory _data) public {
+        data = new uint256[](_data.length);
+        for (uint256 i = 0; i < _data.length; i++) {
             data[i] = _data[i];
         }
     }
@@ -26,9 +27,9 @@ contract UintUtils {
      * @dev Returns minimal element in array
      * @return uint
      */
-    function min() public view returns (uint) {
-        uint minimal = data[0];
-        for (uint i; i < data.length; i++) {
+    function min() public view returns (uint256) {
+        uint256 minimal = data[0];
+        for (uint256 i; i < data.length; i++) {
             if (data[i] < minimal) {
                 minimal = data[i];
             }
@@ -40,9 +41,9 @@ contract UintUtils {
      * @dev Returns minimal element's index
      * @return uint
      */
-    function imin() public view returns (uint) {
-        uint minimal = 0;
-        for (uint i; i < data.length; i++) {
+    function imin() public view returns (uint256) {
+        uint256 minimal = 0;
+        for (uint256 i; i < data.length; i++) {
             if (data[i] < data[minimal]) {
                 minimal = i;
             }
@@ -54,9 +55,9 @@ contract UintUtils {
      * @dev Returns maximal element in array
      * @return uint
      */
-    function max() public view returns (uint) {
-        uint maximal = data[0];
-        for (uint i; i < data.length; i++) {
+    function max() public view returns (uint256) {
+        uint256 maximal = data[0];
+        for (uint256 i; i < data.length; i++) {
             if (data[i] > maximal) {
                 maximal = data[i];
             }
@@ -68,9 +69,9 @@ contract UintUtils {
      * @dev Returns maximal element's index
      * @return uint
      */
-    function imax() public view returns (uint) {
-        uint maximal = 0;
-        for (uint i; i < data.length; i++) {
+    function imax() public view returns (uint256) {
+        uint256 maximal = 0;
+        for (uint256 i; i < data.length; i++) {
             if (data[i] > data[maximal]) {
                 maximal = i;
             }
@@ -82,9 +83,9 @@ contract UintUtils {
      * @dev Compute sum of all elements
      * @return uint
      */
-    function sum() public view returns (uint) {
-        uint S;
-        for (uint i; i < data.length; i++) {
+    function sum() public view returns (uint256) {
+        uint256 S;
+        for (uint256 i; i < data.length; i++) {
             S += data[i];
         }
         return S;
@@ -94,7 +95,7 @@ contract UintUtils {
      * @dev assign new array pointer from _data
      * @param _data is array to assign
      */
-    function set(uint[] memory _data) public {
+    function set(uint256[] memory _data) public {
         data = _data;
     }
 
@@ -102,23 +103,23 @@ contract UintUtils {
      * @dev Get the contents of array
      * @return uint[]
      */
-    function get() public view returns (uint[] memory) {
+    function get() public view returns (uint256[] memory) {
         return data;
     }
 
-    function at(uint i) public view returns (uint) {
+    function at(uint256 i) public view returns (uint256) {
         return data[i];
     }
 
-    function sort_item(uint pos) internal returns (bool) {
-        uint w_min = pos;
-        for (uint i = pos; i < data.length; i++) {
+    function sort_item(uint256 pos) internal returns (bool) {
+        uint256 w_min = pos;
+        for (uint256 i = pos; i < data.length; i++) {
             if (data[i] < data[w_min]) {
                 w_min = i;
             }
         }
         if (w_min == pos) return false;
-        uint tmp = data[pos];
+        uint256 tmp = data[pos];
         data[pos] = data[w_min];
         data[w_min] = tmp;
         return true;
@@ -128,7 +129,7 @@ contract UintUtils {
      * @dev Sort the array
      */
     function sort() public {
-        for (uint i = 0; i < data.length - 1; i++) {
+        for (uint256 i = 0; i < data.length - 1; i++) {
             sort_item(i);
         }
     }
@@ -137,8 +138,9 @@ contract UintUtils {
 // Int Array
 
 contract IntUtils {
-    int[] private data;
-    function IntArray(int[] memory _data) public {
+    int256[] private data;
+
+    function IntArray(int256[] memory _data) public {
         data = _data;
     }
 
@@ -146,9 +148,9 @@ contract IntUtils {
      * @dev Returns minimal element in array
      * @return int
      */
-    function min() public view returns (int) {
-        int minimal = data[0];
-        for (uint i; i < data.length; i++) {
+    function min() public view returns (int256) {
+        int256 minimal = data[0];
+        for (uint256 i; i < data.length; i++) {
             if (data[i] < minimal) {
                 minimal = data[i];
             }
@@ -160,9 +162,9 @@ contract IntUtils {
      * @dev Returns minimal element's index
      * @return uint
      */
-    function imin() public view returns (uint) {
-        uint minimal = 0;
-        for (uint i; i < data.length; i++) {
+    function imin() public view returns (uint256) {
+        uint256 minimal = 0;
+        for (uint256 i; i < data.length; i++) {
             if (data[i] < data[minimal]) {
                 minimal = i;
             }
@@ -174,9 +176,9 @@ contract IntUtils {
      * @dev Returns maximal element in array
      * @return int
      */
-    function max() public view returns (int) {
-        int maximal = data[0];
-        for (uint i; i < data.length; i++) {
+    function max() public view returns (int256) {
+        int256 maximal = data[0];
+        for (uint256 i; i < data.length; i++) {
             if (data[i] > maximal) {
                 maximal = data[i];
             }
@@ -188,9 +190,9 @@ contract IntUtils {
      * @dev Returns maximal element's index
      * @return uint
      */
-    function imax() public view returns (uint) {
-        uint maximal = 0;
-        for (uint i; i < data.length; i++) {
+    function imax() public view returns (uint256) {
+        uint256 maximal = 0;
+        for (uint256 i; i < data.length; i++) {
             if (data[i] > data[maximal]) {
                 maximal = i;
             }
@@ -202,35 +204,35 @@ contract IntUtils {
      * @dev Compute sum of all elements
      * @return int
      */
-    function sum() public view returns (int) {
-        int S;
-        for (uint i; i < data.length; i++) {
+    function sum() public view returns (int256) {
+        int256 S;
+        for (uint256 i; i < data.length; i++) {
             S += data[i];
         }
         return S;
     }
 
-    function set(int[] memory _data) public {
+    function set(int256[] memory _data) public {
         data = _data;
     }
 
-    function get() public view returns (int[] memory) {
+    function get() public view returns (int256[] memory) {
         return data;
     }
 
-    function at(uint i) public view returns (int) {
+    function at(uint256 i) public view returns (int256) {
         return data[i];
     }
 
-    function sort_item(uint pos) internal returns (bool) {
-        uint w_min = pos;
-        for (uint i = pos; i < data.length; i++) {
+    function sort_item(uint256 pos) internal returns (bool) {
+        uint256 w_min = pos;
+        for (uint256 i = pos; i < data.length; i++) {
             if (data[i] < data[w_min]) {
                 w_min = i;
             }
         }
         if (w_min == pos) return false;
-        int tmp = data[pos];
+        int256 tmp = data[pos];
         data[pos] = data[w_min];
         data[w_min] = tmp;
         return true;
@@ -240,7 +242,7 @@ contract IntUtils {
      * @dev Sort the array
      */
     function sort() public {
-        for (uint i = 0; i < data.length - 1; i++) {
+        for (uint256 i = 0; i < data.length - 1; i++) {
             sort_item(i);
         }
     }
